% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spps_final_estimators.R
\name{atemod}
\alias{atemod}
\title{Estimates the avarage treatment effect (ATE)  using the SPPS model}
\usage{
atemod(x, a, y, phi = expit, dphi = dexpit, phiinv = expitinv,
  estim.method = "ML", maxit = 25, type = 1)
}
\arguments{
\item{x}{A matrix of covariates of dimension n x p}

\item{a}{A vector a of 0s and 1s of dimension n}

\item{y}{A vector of responses or outcomes}

\item{phi}{A function that maps the real line onto the interval (0,1)}

\item{dphi}{The derivative of \code{phi}}

\item{phiinv}{The inverse of \code{phi}}

\item{estim.method}{The estimation method to be used for estimating the parameters "ML" (default) or "LS"}

\item{maxit}{The maximum number of iterations to be performed.}

\item{type}{The type of IPW estimator to be used: type 1 , type 2 and type 3  combine SPPS with, respectively, IPW1, IPW2 and IPW3 in Lunceford and Davidian(2004)}
}
\value{
a list containing the ATE and the estimators of \code{beta}, \code{epsilon} and \code{delta}
}
\description{
Estimates the avarage treatment effect (ATE)  using the SPPS model
}
