% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spps_final_estimators.R
\name{estibeta}
\alias{estibeta}
\title{Estimates beta, knowing epsilon and delta.}
\usage{
estibeta(x, a, epsilon, delta, estim.method = "ML", phi = expit,
  dphi = dexpit, phiinv = expitinv, maxit = 3)
}
\arguments{
\item{x}{A matrix of covariates of dimension n x p}

\item{a}{A vector a of 0s and 1s of dimension n}

\item{epsilon}{A real number between 0 and 1 such that \code{epsilon}+\code{delta}<1}

\item{delta}{A real number between 0 and 1}

\item{estim.method}{The estimation method to be used for estimating the parameters "ML" (default) or "LS"}

\item{phi}{A function that maps the real line onto the interval (0,1)}

\item{dphi}{The derivative of \code{phi}}

\item{phiinv}{The inverse of \code{phi}}

\item{maxit}{The maximum number of iterations to be performed.}
}
\value{
a real number between 0 and 1
}
\description{
Estimates beta, knowing epsilon and delta.
}
