% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spps_ML_estimators.R
\name{estidelML}
\alias{estidelML}
\title{Estimates delta, knowing epsilon and  eta=X%*%beta, by maximum likelihood.}
\usage{
estidelML(a, eta, epsilon, phi = expit)
}
\arguments{
\item{a}{A Vector a of 0s and 1s of dimension n}

\item{eta}{A vector of linear predictors of dimension n}

\item{epsilon}{A real number between 0 and 1}

\item{phi}{A function that maps the real line onto the interval (0,1)}
}
\value{
a real number between 0 and 1
}
\description{
Estimates delta, knowing epsilon and  eta=X%*%beta, by maximum likelihood.
}
