% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spps_ML_estimators.R
\name{estiepsML}
\alias{estiepsML}
\title{Estimates epsilon, knowing delta and  eta=X%*%beta, by maximum likelihood.}
\usage{
estiepsML(a, eta, delta, phi = expit)
}
\arguments{
\item{a}{A vector a of 0s and 1s of dimension n}

\item{eta}{A vector of linear predictors of dimension n}

\item{delta}{A real number between 0 and 1}

\item{phi}{A function that maps the real line onto the interval (0,1)}
}
\value{
a real number between 0 and 1
}
\description{
Estimates epsilon, knowing delta and  eta=X%*%beta, by maximum likelihood.
}
